ref: move files and move message passing to new file
