+++ packages/i18n/src/useTranslation/getExtensionMessage.ts
